pyproject.toml
[project]
name = "agent-demo"
version = "0.1.0"
description = "Traffic Intelligence Agent Engineering Demo"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gradio>=6.3.0",
    "httpx>=0.28.1",
    "langchain-openai>=0.3.1",
    "langgraph>=0.2.66",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
